introduce StateTerminationClass/Type and assign type to every state termination
* removes --silent-klee-assume (instead don't use "early" in --write-tests)
* removes --dump-states-on-halt (instead (don't) use "early" in --write-tests)
* introduces --write-tests=none|exit,early,error,user-error,execution-error

All causes for state termination are listed in StateTerminationType and
categorised according to StateTerminationClass. The StateTerminationClass in
combination with --write-tests is used to decide wether a test should be
written. For three early termination cases no tests are generated (solver
failure/timeout, klee_silent_exit() and merged states).

-only-output-states-covering-new filters states with "exit" and "early"
termination class.

Example: --write-tests=error,execution-error only generates tests when bugs are
found in the program or the execution hits a bug/unimplemented feature in KLEE.